pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "fastapi>=0.81.0",
    "uvicorn>=0.18.3",
    "pytest>=7.1.3",
    "requests>=2.28.1",
    "coverage>=6.4.4",
    "python-dateutil>=2.8.2",
    "SQLAlchemy>=1.4.41",
    "alembic>=1.8.1",
    "psycopg2-binary>=2.9.3",
    "orjson>=3.8.0",
    "databases>=0.6.1",
    "asyncpg>=0.26.0",
    "httpx>=0.23.0",
    "anyio>=3.6.1",
    "trio>=0.21.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
